Fixed Crash With Floating-Point Storage Counts
~ Temporarily changed StorageDTO resourceCapacity and resourceCount to float ... doesnt fix the underlying server issue of storing floats on storages to begin with, but it will allow us to continue with testing the unity client and get stuff spawning in

+++ unity-client/Assets/Scripts/DTO/StorageDTO.cs
@@ -1,6 +1,6 @@
 public class StorageDTO
 {
     public ResourceType resourceType;
-    public int resourceCapacity;
-    public int resourceCount;
+    public float resourceCapacity;
+    public float resourceCount;
 }
